code style fixed
gossip network commit 1

gossip network commit 2

gossip network commit 3

gossip network commit 4

fixed some compile errors

cluster startup

add file cluster.cpp

compile problem

encode unit test

fix some bugs

cluster meet

fix compile error

gossip message handling

fix compile error

fix some bugs

+++ src/tendisplus/utils/string.cpp
@@ -264,7 +264,7 @@ uint64_t getCurThreadId() {
     // TODO(vinchen): the performance is?
     oss << std::this_thread::get_id();
     std::string stid = oss.str();
-    unsigned long long tid = std::stoull(stid);
+    uint64_t tid = std::stoull(stid);
 
     return tid;
 }
